discordBot-28-cumbres/requirements.txt
discord
python-dotenv
requests
pymongo
apscheduler